plic works
