+++ kernel/trap.c
@@ -198,19 +198,19 @@ devintr()
 
     // irq indicates which device interrupted.
     #ifdef QEMU
-    // int irq = plic_claim();
-    // if(irq == UART0_IRQ){
-    //   uartintr();
-    // } else if(irq == VIRTIO0_IRQ){
-    //   virtio_disk_intr();
-    // } else if(irq){
-    //   printf("unexpected interrupt irq=%d\n", irq);
-    // }
-    // // the PLIC allows each device to raise at most one
-    // // interrupt at a time; tell the PLIC the device is
-    // // now allowed to interrupt again.
-    // if(irq)
-    //   plic_complete(irq);
+    int irq = plic_claim();
+    if(irq == UART0_IRQ){
+      uartintr();
+    } else if(irq == VIRTIO0_IRQ){
+      disk_intr();
+    } else if(irq){
+      printf("unexpected interrupt irq=%d\n", irq);
+    }
+    // the PLIC allows each device to raise at most one
+    // interrupt at a time; tell the PLIC the device is
+    // now allowed to interrupt again.
+    if(irq)
+      plic_complete(irq);
     #endif
 
     return 1;
