+++ kernel/virtio_disk.c
@@ -14,7 +14,6 @@
 #include "include/memlayout.h"
 #include "include/spinlock.h"
 #include "include/sleeplock.h"
-#include "include/fs.h"
 #include "include/buf.h"
 #include "include/virtio.h"
 
@@ -49,7 +48,7 @@ static struct disk {
 } __attribute__ ((aligned (PGSIZE))) disk;
 
 void
-virtio_disk_init(void)
+disk_init(void)
 {
   uint32 status = 0;
 
@@ -171,7 +170,7 @@ alloc3_desc(int *idx)
 void
 virtio_disk_rw(struct buf *b, int write)
 {
-  uint64 sector = b->blockno * (BSIZE / 512);
+  uint64 sector = b->sectorno;
 
   acquire(&disk.vdisk_lock);
 
@@ -252,7 +251,7 @@ virtio_disk_rw(struct buf *b, int write)
 }
 
 void
-virtio_disk_intr()
+disk_intr()
 {
   acquire(&disk.vdisk_lock);
 
@@ -270,4 +269,16 @@ virtio_disk_intr()
   *R(VIRTIO_MMIO_INTERRUPT_ACK) = *R(VIRTIO_MMIO_INTERRUPT_STATUS) & 0x3;
 
   release(&disk.vdisk_lock);
+}
+
+void 
+disk_read(struct buf *b) {
+	virtio_disk_rw(b, 0);
+}
+
+/* 
+ * Write content into disk 
+ */
+void disk_write(struct buf *b) {
+	virtio_disk_rw(b, 1);
 }
